backend/pyproject.toml
[tool.poetry]
name = "task-mgmt-backend"
version = "0.1.0"
description = "Task Management System backend"
authors = ["Hassan Abouelela <hassan@hassanamr.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
fastapi = "0.117.0"
uvicorn = { version = "0.36.0", extras = ["standard"] }
sqlalchemy = "2.0.43"
psycopg = { version = "3.2.10", extras = ["binary"] }
pydantic-settings = "2.10.1"
python-dotenv = "1.1.1"
agno = "2.0.7"
pyjwt = "2.10.1"
passlib = { version = "1.7.4", extras = ["bcrypt"] }

[tool.poetry.group.dev.dependencies]
pytest = "8.4.2"
httpx = "0.27.2"
ruff = "0.5.7"
black = "25.9.0"
isort = "6.0.1"
pre-commit = "4.3.0"
atlas-provider-sqlalchemy = "0.4.0"

[tool.ruff]
line-length = 120

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = "."
